requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
docopt==0.6.2
idna==2.10
Pillow==10.0.0
pipreqs==0.4.10
pycparser==2.20
pylightxl==1.53
PyNaCl==1.4.0
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.2
urllib3==1.26.3
XlsxWriter==1.3.7
yarg==0.1.9
